// src/app/models/SasType.ts
export enum SasType{
    TCAE_JOB_BANK = "TCAE_JOB_BANK",
    TEAP_JOB_BANK = "TEAP_JOB_BANK",
    TEDN_JOB_BANK = "TEDN_JOB_BANK",
    TEDS_JOB_BANK = "TEDS_JOB_BANK",
    TEL_JOB_BANK = "TEL_JOB_BANK",
    TEMN_JOB_BANK = "TEMN_JOB_BANK",
    TERD_JOB_BANK = "TERD_JOB_BANK ",
    TER_JOB_BANK = "TER_JOB_BANK",
    TF_JOB_BANK = "TF_JOB_BANK ",
    NURSE_JOB_BANK = "NURSE_JOB_BANK",
    FISIO_JOB_BANK = "FISIO_JOB_BANK",
    SPEECH_THERAPIST_JOB_BANK = "SPEECH_THERAPIST_JOB_BANK",
    OCCUPATIONAL_THERAPY_JOB_BANK = "OCCUPATIONAL_THERAPY_JOB_BANK",
    NURSE_FAMILY_JOB_BANK = "NURSE_FAMILY_JOB_BANK",
    NURSE_GYNECOLOGY_JOB_BANK = "NURSE_GYNECOLOGY_JOB_BANK",
    NURSE_PEDIATRICIAN_JOB_BANK = "NURSE_PEDIATRICIAN_JOB_BANK",
    NURSE_MENTAL_HEALTH_JOB_BANK = "NURSE_MENTAL_HEALTH_JOB_BANK",
    NURSE_WORK_JOB_BANK = "NURSE_WORK_JOB_BANK",
    NURSE_CRITICS_SPECIFIC_JOB_BANK = "NURSE_CRITICS_SPECIFIC_JOB_BANK",
    NURSE_DIALYSIS_SPECIFIC_JOB_BANK = "NURSE_DIALYSIS_SPECIFIC_JOB_BANK",
    NURSE_MENTAL_HEALTH_SPECIFIC_JOB_BANK = " NURSE_MENTAL_HEALTH_SPECIFIC_JOB_BANK",
    NURSE_NEONATES_SPECIFIC_JOB_BANK = "NURSE_NEONATES_SPECIFIC_JOB_BANK",
    NURSE_NUCLEAR_SPECIFIC_JOB_BANK = "NURSE_NUCLEAR_SPECIFIC_JOB_BANK",
    NURSE_SURGERY_ROOM_SPECIFIC_JOB_BANK = "NURSE_SURGERY_ROOM_SPECIFIC_JOB_BANK",
    NURSE_EXAM = "NURSE_EXAM",
    TCAE_EXAM = "TCAE_EXAM"
}